requirements.txt
Aspose.Slides==23.1.0